snapconnect-backend/app/requirements.txt
# Absolute minimal dependencies to fix the Render build issue.
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
# openai==1.10.0  # Temporarily disabled - may have compilation dependencies
# pinecone-client==3.0.0  # Temporarily disabled due to build issues
python-multipart==0.0.7
# httpx==0.26.0  # Temporarily disabled - openai dependency
# For image embeddings (optional, comment out if not needed initially)
# transformers==4.35.2
# PyPI wheels for torch is CPU-only by default; plain version is fine
# torch==2.2.0

# New dependencies
httpx==0.27.0
openai==1.30.1
pinecone-client==4.1.0
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7

# Image stack removed for hosted OpenAI embeddings

# Image / Vision stack for CLIP embeddings
# pillow==10.3.0
# transformers==4.41.1 